fix for bug related to pytest 7.0.0, restructured `elephant/test`, helper functions were moved to `testutils`, thus fixing errors when running pytest with  `--import-mode=importlib`
