+++ elephant/test/test_unitary_event_analysis.py
@@ -14,7 +14,7 @@ import quantities as pq
 from numpy.testing import assert_array_equal
 
 import elephant.unitary_event_analysis as ue
-from elephant.test.download import download, ELEPHANT_TMP_DIR
+from elephant.testutils.download import download, ELEPHANT_TMP_DIR
 from numpy.testing import assert_array_almost_equal
 from elephant.spike_train_generation import StationaryPoissonProcess
